+++ src/main/java/com/workoutbuilder/enterprise/controller/WorkoutBuilderController.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,24 +14,14 @@ import java.util.List;
  * Controller class responsible for handling requests related to workouts and exercises.
  */
 @Controller
+@RequestMapping("/api/exercise")
 public class WorkoutBuilderController {
 
-    @Autowired
-    IExerciseService exerciseService;
+    private final IExerciseService exerciseService;
 
-    /**
-     * Maps the root URL ("/") to the index page.
-     *
-     * @param model Model object for UI rendering.
-     * @return name of the index view.
-     */
-    @GetMapping("/")
-    public String index(Model model) {
-        Exercise exercise = new Exercise();
-        model.addAttribute("exercise", exercise);
-        List<Exercise> exercises = exerciseService.findAll();
-        model.addAttribute("exercises", exercises);
-        return "/index";
+    @Autowired
+    public WorkoutBuilderController(IExerciseService exerciseService) {
+        this.exerciseService = exerciseService;
     }
 
     /**
@@ -40,7 +29,7 @@ public class WorkoutBuilderController {
      *
      * @return List of exercises.
      */
-    @GetMapping("api/exercise")
+    @GetMapping("/")
     @ResponseBody
     public List<Exercise> fetchAllExercises(){
         return exerciseService.findAll();
@@ -52,7 +41,7 @@ public class WorkoutBuilderController {
      * @param id ID of the exercise.
      * @return Exercise with the given ID.
      */
-    @GetMapping("api/exercise/{id}")
+    @GetMapping("/{id}")
     @ResponseBody
     public Exercise fetchExerciseById(@PathVariable("id") int id){
         return exerciseService.findById(id);
@@ -64,7 +53,7 @@ public class WorkoutBuilderController {
      * @param exercise Exercise object to be created.
      * @return Newly created exercise.
      */
-    @PostMapping(value = "api/exercise", consumes = "application/json", produces = "application/json")
+    @PostMapping(value = "/", consumes = "application/json", produces = "application/json")
     @ResponseBody
     public Exercise createExercise(@RequestBody Exercise exercise) {
         Exercise newExercise = null;
@@ -82,8 +71,8 @@ public class WorkoutBuilderController {
      * @param exercise Exercise object to be saved.
      * @return Redirection to the index page.
      */
-    @RequestMapping("api/saveExercise")
-    public String saveExercise(Exercise exercise) {
+    @PutMapping("/updateExercise")
+    public String updateExercise(Exercise exercise) {
         try {
             exerciseService.saveExercise(exercise);
         } catch (Exception e) {
@@ -99,13 +88,13 @@ public class WorkoutBuilderController {
      * @param id ID of the exercise to be deleted.
      * @return ResponseEntity with either HTTP OK status or HTTP Internal Server Error status.
      */
-    @DeleteMapping("api/exercise/{id}")
-    public ResponseEntity deleteExercise(@PathVariable("id") int id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteExercise(@PathVariable("id") int id) {
         try{
             exerciseService.deleteExercise(id);
-            return new ResponseEntity(HttpStatus.OK);
+            return new ResponseEntity<>("Entry successfully deleted.", HttpStatus.OK);
         } catch (Exception e) {
-            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>("Entry delete request FAILED.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
